requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
bcrypt
PyJWT
pydantic-settings
email-validator
python-multipart
